fix type of creationDate in HabitViewHolder replace and implemented the logic of getting data from server and sending them to local db in HabitRepository

+++ app/src/main/java/com/example/newapppp/data/remote/habit/HabitDone.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class HabitDone(
     @SerialName(value = "date")
-    val creationDate: Int,
+    val creationDate: Long,
     @SerialName(value = "habit_uid")
     val id: String
 )
